Had some complications
Tried to save the file as they did in the book, and changed the
index.html file, but it did not like that. I even had to change the
index.html file back in order to delete the file.

So, I don’t know how to save .js scripts from the book’s examples at
this point