Dog Years
This was a bit tricky, I changed the .js file and the index.html file
to ‘dogYears’ for both. It ran the same with all four up an running.
However I am curious to see if I delete the original index and .js
files if it will run the same. Otherwise, I don’t know how we can
complete the assignment without having a .js script and a html file for
each problem.