sliceOfPiePartTwo with some fixes for part One
I needed to really focus on the English for both parts. There shouldn’t
be any confusion as to what each variable is for, where is belongs and
how it is worked out to come to the result. I also needed to make sure
that the results to print were exactly what the instructions asked for.