Ternary's!!!! I needed a ternary to finish the code
Okay…now we have the tools we need to find out whether or not you need
another air conditioner or not. The code is clean, but long. Now that I
have time, I want to shorten it up a bit and maybe put some parameters
(so to speak) that will allow us to skip past code (like if you only
have 1 or 2 rooms and just 1 bath). Good time management for me!!!