Cleaned up the console.log responses
I kept what was useful, eliminated what didn’t make sense (such as
negative results) and stayed focused on the end result….do they need
another air conditioner or not. I will use the ‘test’ sheet to
experiment with nesting the bedroom and bathroom problem. Time is of
the essence.